backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
PyJWT==1.7.1
sqlparse==0.5.3
tzdata==2025.1
